Fixed HDG load balance weighting method: Changed ElemHDGSides to REAL to account for periodic side (counted with factor 0.5) and removed Dirichlet sides from been considered

+++ src/loadbalance/loadbalance_vars.f90
@@ -78,7 +78,7 @@ REAL                                :: ElemTimePartTot  ! Total time spent for p
 REAL                                :: ElemTimeFieldTot ! Total time spent for field routines (all procs)
 REAL                                :: ElemTimePart     ! Time spent for particle routines
 REAL                                :: ElemTimeField    ! Time spent for field routines
-INTEGER,ALLOCATABLE                 :: ElemHDGSides(:)  ! number of master sides for the HDG solver for each element
+REAL,ALLOCATABLE                    :: ElemHDGSides(:)  ! number of master sides for the HDG solver for each element
 INTEGER                             :: TotalHDGSides    ! total number of master sides for the HDG solver over all local elements
 REAL,ALLOCATABLE                    :: ElemGlobalTime(:)
 INTEGER(KIND=8),ALLOCATABLE         :: nPartsPerElem(:)
